Renamed GSurv package
