+++ pkg/boot/gsurv/gsurv.go
@@ -1,4 +1,4 @@
-package mudp
+package gsurv
 
 import (
 	"context"
