+++ pkg/boot/gsurv/gsurv_test.go
@@ -1,4 +1,4 @@
-package mudp
+package gsurv
 
 import (
 	"context"
